# Implement a fast path for integer parsing